# internal/platform/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: Backend CRM Piposmart
  version: 0.15.4-sprint-14f-identity
  description: API CRM Piposmart dengan Partner Type, Partner (Encrypted & Masked Bank Account), PIC Assignment, Call Interactions, Lead Referrals, Subscription Order, Reconciliation Sprint 11, Partner Commission Sprint 12, penambahan route trash/unscoped owner-outlet dan rekap status langganan outlet Sprint 14b/14c, histori import persisten dengan viewer/download file asli Sprint 14 history, route /all dan /all-deleted lintas modul untuk kebutuhan frontend mengambil full dataset tanpa pagination, serta penambahan manajemen akun Admin dan Supervisor beserta reset password lintas role pada Sprint 14f.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/closings:
    get:
      summary: List closing penjualan
      operationId: listSalesClosings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_RECONCILIATION, CONFIRMED, REJECTED]
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closed_from
          in: query
          schema:
            type: string
            format: date
        - name: closed_to
          in: query
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List closing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/leads/{lead_id}/closings:
    post:
      summary: Membuat closing untuk lead Sales
      operationId: createLeadClosing
      description: Membuat sales closing, snapshot catalog, dan remark score 3 dalam satu transaksi database.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClosingRequest"
      responses:
        "201":
          description: Closing dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          description: Lead sudah memiliki closing pending/confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/closings/{closing_id}:
    get:
      summary: Detail closing
      operationId: getSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      responses:
        "200":
          description: Detail closing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    delete:
      summary: Soft delete closing
      operationId: deleteSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      responses:
        "200":
          description: Closing soft deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/{closing_id}/confirm:
    post:
      summary: Confirm closing
      operationId: confirmSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClosingStatusRequest"
      responses:
        "200":
          description: Closing confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/{closing_id}/reject:
    post:
      summary: Reject closing
      operationId: rejectSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClosingStatusRequest"
      responses:
        "200":
          description: Closing rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/{closing_id}/restore:
    patch:
      summary: Restore closing soft-deleted
      operationId: restoreSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      responses:
        "200":
          description: Closing restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/{closing_id}/force:
    delete:
      summary: Hard delete closing
      operationId: forceDeleteSalesClosing
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClosingID"
      responses:
        "200":
          description: Closing hard deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/bulk:
    delete:
      summary: Bulk soft delete closing
      operationId: bulkDeleteSalesClosing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIDRequest"
      responses:
        "200":
          description: Bulk soft delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/bulk/restore:
    patch:
      summary: Bulk restore closing
      operationId: bulkRestoreSalesClosing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIDRequest"
      responses:
        "200":
          description: Bulk restore result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/bulk/force:
    delete:
      summary: Bulk hard delete closing
      operationId: bulkForceDeleteSalesClosing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkIDRequest"
      responses:
        "200":
          description: Bulk hard delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /api/v1/closings/trash:
    get:
      summary: List closing yang soft deleted
      operationId: listSalesClosingsTrash
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_RECONCILIATION, CONFIRMED, REJECTED]
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closed_from
          in: query
          schema:
            type: string
            format: date
        - name: closed_to
          in: query
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List closing dalam trash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/unscoped:
    get:
      summary: List semua closing termasuk soft deleted
      operationId: listSalesClosingsUnscoped
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_RECONCILIATION, CONFIRMED, REJECTED]
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closed_from
          in: query
          schema:
            type: string
            format: date
        - name: closed_to
          in: query
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: List closing aktif + terhapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /api/v1/closings/all:
    get:
      summary: List semua closing tanpa pagination
      operationId: listSalesClosingsAll
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_RECONCILIATION, CONFIRMED, REJECTED]
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closed_from
          in: query
          schema:
            type: string
            format: date
        - name: closed_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List semua closing aktif
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/closings/all-deleted:
    get:
      summary: List semua closing termasuk soft deleted tanpa pagination
      operationId: listSalesClosingsAllDeleted
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SearchQuery"
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING_RECONCILIATION, CONFIRMED, REJECTED]
        - name: sales_id
          in: query
          schema:
            type: integer
            format: int64
        - name: supervisor_id
          in: query
          schema:
            type: integer
            format: int64
        - name: closed_from
          in: query
          schema:
            type: string
            format: date
        - name: closed_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List semua closing aktif + terhapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    LeadID:
      name: lead_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ClosingID:
      name: closing_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    SearchQuery:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Search global untuk kode, nama, telepon, brand, kota, dan provinsi.
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  schemas:
...
    BulkIDRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
          example: [1, 2, 3]
...
    CreateClosingRequest:
      type: object
      required: [plan_id]
      properties:
        plan_id:
          type: integer
          format: int64
        promotion_id:
          type: integer
          format: int64
          nullable: true
        discount_amount:
          type: string
          example: "0.00"
        unique_transfer_code:
          type: integer
          minimum: 0
          maximum: 999
        closed_at:
          type: string
          format: date-time
        interaction_type:
          type: string
          enum: [CALL, CHAT]
        contact_name:
          type: string
        contact_phone:
          type: string
        customer_response:
          type: string
        note:
          type: string
    UpdateClosingStatusRequest:
      type: object
      properties:
        note:
          type: string
        reason:
          type: string
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          additionalProperties: true
        meta:
          type: object
          required: [request_id]
          properties:
            request_id:
              type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true
            request_id:
              type: string
...